Context-based texture caching.
Adds an additional `context' parameter to all of the texture
lookup routines.  When searching, we only consider textures
equivalent if their creation and use is from the same context.
This prevents us from reusing textures in which the creation and
use contexts differ.

This is useful for client applications that want to use Tuvok to
render multiple times using a single process.  The application can
establish N contexts and tell Tuvok to render N times, changing
view parameters (or whatever) each time.  Without this code, we'd
crash trying to share data between the renderings.