Switch to a size_t for the component count.
Prevents a warning about a narrowing conversion on Win32, at
least.  This required temporarily widening the variable on
read/write, because I didn't want to change the file format again.

Note that this means Tuvok can only load data with 4.2 billion or
fewer components.  Let me know when this becomes an issue...