Add jpeg to our third party libs.
This is actually a version from the GDCM library, which is
essentially IJG's JPEG reader plus some patches to support
lossless JPEGs, plus some other small modifications.  I've made a
few small modifications of my own so that the library works better
with Tuvok, as well.

There is a compile time constant which sets the number of bits per
component within `jconfig.h'.  I've set this to 16 to allow
reading the test image I have.  However, 8 and 12 bit lossless
JPEGs supposedly exist as well.  The solution to this problem that
others use is to compile the library 3 times, once for each bit
size, mangling the symbols a tiny bit each time.  We might have to
do the same eventually.