Frontend implementation for custom jpeglib-based JPEG loading.
The wrapper can load a JPEG from either a file, or an existing
in-memory buffer.  The latter causes a large amount of pain; see
the implementation for details.

Basically one creates a tuvok::JPEG object, giving it a file or a
chunk of memory which holds a JPEG.  The header is parsed during
initialization, and a `valid' method details whether or not we
could successfully read the header.  If so, a `data' method
decodes the image and returns a pointer to it.

Thanks for the help, QImage.