Remove jmemansi.c.
Yet another overcomplicated memory manager.  Based around the idea
that if you don't have enough memory, it would create a file to
back the memory.  Most of us just call this `swapping'.

What's a large JPEG anyway?  5 meg?  If a user can't fit 5 meg in
core, I don't think Tuvok will deliver the performance we hype.

Or an image, for that matter.