Support skipping some bytes before reading a JPEG payload.
Adds an offset parameter to allow us to skip some number of bytes
in a file before attempting to read a JPEG in it.  Needed/useful
in the case that a JPEG is embedded in another file, such as a
DICOM.

We assume that the JPEG begins at the given offset and continues
until the end of the file, which wouldn't strictly be correct in
the DICOM case.  In practice it ends up being true almost always,
but in those other cases we require more memory than is strictly
necessary.