Handle GL out of memory conditions gracefully.
glTexImage?D can fail due to the driver's inability to obtain
memory.  For the 3D case, we started throwing an exception when
this occurs.  This adds code to handle the exception, attempting
to delete more bricks than we had previously thought necessary.

It proved a bit difficult to unify the 'replace a 3d texture' path
with the 'just delete bricks until the new texture can fit' path.
This implementation uses a wrapper function which calls the old
'Get3DTexture' (i.e. both paths) and catches the exception above
it.