Add new `Quantize' header.
Despite the name, it only contains templatized minmax code right
now.  It will expand to include most of the AbstrConverter
quantization code.  Most of this will get moved into an
implementation file at that point.

The implementation is policy based.  Currently the two policies
are for data access and progress notification.  For data access, a
policy is included for both C++ io streams and Basics'
LargeRAWFile, which is typically faster.  For progress
notification, a Null indicator is available as well as one which
outputs Tuvok debug messages.