Add pathway to set a 1D TF from an external source.
Apps that don't use UVF need some way to set the transfer function.
Grabbing the current TF out of the renderer and modifying it is
insufficient, because changes in size to the TF won't get
reflected in the GPU texture.

This adds a new accessor in the renderer to set the 1D transfer
function, which handles deallocating the previous TF data and
reallocating data of the correct size.