Support a data range query.
Adds some new accessors to pull out the range of values in a
dataset.  Currently unimplemented for UVF, but we'd never call it
when accessing a UVF file.

In-memory datasets are given the option of explicitly setting the
range.  If they don't, the range will be automatically calculated
when they upload data.

This also cleans up some implementation code for in-memory
datasets, by using some template code to avoid duplicating some
code as datasets are copied into Tuvok.