SBVR: Support alternate TF scaling method.
Allow scaling given a bias amount and a scaling factor.

This requires setting an additional uniform (the bias) and
resetting our scaling factor for the new method.  It also requires
using a modified shader...

Unfortunately a client app needs to define BIAS_SCALE in the
shader to get this to work, and they can't really switch around
methods at runtime.  That's not so terrible, because a client app
would copy Tuvok's shaders into its distribution anyway, but is
still rather annoying.  Doing it right requires that ethereal
shader factory we keep talking about ...