GPU-MM functionality for external DS's.
Create 'AddDataset' and make 'FreeAssociatedTextures' public.
This allows the renderer to add datasets, plus cleanup 3D textures
when new data are uploaded.  External datasets are a special case
in this respect: data often change at runtime, and when that
occurs any old textures are useless -- we're guaranteed they'll
never be used again.