Propagate debug message settings.
If we add a new channel that has messages enabled, the multiplexer
must get called for messages regardless of whether it was getting
called for them before.

To allow channels which are multiplexed to have different states,
we'll filter events per-debugger.  Note that changing the debug
states of a multiplexer implicitly changes the debug states of all
debug outputs that it multiplexes, so any uniqueness among
multiplexed debug outs is lost when the state is changed by the
user.

Still, this enables the useful use case of enabling ConsoleOut
messages without QTOut messages, at least until one fiddles with
the UI.  This is convenient because QTOut is slow as hell, but the
performance impact for the non-GUI outs isn't so bad.