Add an assert to catch a common mistake.
If some brick metadata doesn't get set correctly, it ends up
visible in the slice generating code by creating an infinite loop.
This adds some logic so that one will notice in debug mode.

I'm a little concerned the preprocessor magic isn't valid.  It may
be the case that tr1 adopted isnan; should verify and see if we
can rely on that.