Fix slice rendering for elongated datasets.
For datasets which had a single dimension which differed from the
other two + that dimension was non-PoT, we'd calculate a bad slice
index along that unique dimension.  The issue was in scaling the
location of the slice index.  Now we convert to the 0-1 space of
the actual data, and do a second scaling to convert to the space of
the texture, all in Render2DView instead of trying to do it in
RenderSlice.