Store/manipulate isovalues `natively'.
That is, AbstrRenderer's isovalue is in the range of the data.
The UI needs integers, so it is responsible for that conversion;
the GL renderer needs data in the range [0,1], so it does that
conversion (that said, a convenience method,
GetNormalizedIsovalue, was added for the latter conversion).

The isovalue defaults to half the data range (as before).