Fix a subtle issue deleting 3D textures.
Deleting an element from the beginning or end of a deque
invalidates all iterators pointing to the deleted element.
However, deleting an element in the middle of a deque invalidates
*all* iterators.

That is, with an 8-element deque, if we delete element 4, we
should not continue searching at element 4 of 7.  If we do, we
could miss values we're looking for, depending on how the deque
is implemented.

The new algorithm, implemented here, completely restarts the
search when a texture gets deleted, thereby making no assumptions
on iterator validity.