Use old window size default.
We want a resize to actually be a resize.  0,0 is a convenient
hint that our client app isn't being very nice, and hasn't given
us a size yet.  We should probably just require a size in the
constructor..

We don't know / can't give a valid number of Y pixels without a
default size.  Just use 1, and rely on the fact that we'll set the
frustum up correctly later.