fixed a bug introduced by Tiago's commit 1446, namely the removal of the filled buffers counter, please do only remove variables if you actually understand what they are god for!!!
OpenGL uses not just one but multiple framebuffers thus copying images only if something changed is not a good idea as that only writes data into one of the framebuffers, I now fixed this adding code to blit the last result into the current framebuffer in every frame, my previous code (removed by Tiago) was more efficient as it only blitted twice (I don't know a tripple buffered GL context).

Also I do not understand why we would need the "displayBufferIsGarbage" var (that I renamed to m_bDisplayIsUninitialized btw.) as it is guaranteed that at first startup Resize is called which also executes a clear color buffer. On my system removing the code that is related to "displayBufferIsGarbage" (namely the additional clear) does not change anything (as expected). So please verify on your system that this code is actually necessary, otherwise remove it.