Undid previous change, r1449 by Jens, that blitted the last results to the framebuffer every frame regardless of whether the last results were correct. It is possible for that buffer to contain invalid information; for instance, when a resize occurs the old FBO will be destroyed and a new one created that will be in an uninitialized state. We don't want to blit that to the display. Secondly, OpenGL's swap buffers function is a bit of a dated misnomer. It doesn't actually specify that a swap occurs. Only that the back buffer information goes to the front buffer some how. From all the systems I've tried it looks like opengl blits the back buffer to the front buffer and so keeps the back buffer in the same state. I haven't tested on windows though, so it wouldn't surprise me if something totally different occurs. After all, the state of the back buffer is technically undefined in the opengl spec... Let me know if things are screwy on windows. Or, if we are uncomfortable with the undefined nature and would prefer taking a small performance hit we can use the multiple render regions path that keeps track of its own back buffer. This would cost us a blit from the current FBO to the last FBO, but should be fully correct and not much more expensive than the blitting to the back buffer that was being done each frame in r1449. If profiling shows that this is noticeably slower than the original code we could also try putting back in the filled buffer counter but this time do it in a way that still allows resizes and the like to work correctly.
Added a ScheduleDisplayClearing() method that sets m_bDisplayIsUninitialized to
true. This needs to be called by the client in certain situations, such as when
changing the number of render regions (i.e. 2x2 to single) or when the
dimensions of the render regions change. Feel free to rename this method to
something better...

Fixed bug in multiple render regions path where garbage was being displayed
when the region was uninitialized. Now the region is cleared to show the
background.

When rendering multiple render regions the gradient background wasn't specific
to each region. Now the background is per region. Moving the horizontal
splitter in 2x2 mode in the old code should make it clear how this was wrong.