Don't block indefinitely in capture mode.
Nice idea, but the entire UI freezes.  Better to just have
multiple paint calls, so the UI can respond to events in between
them.

Also, gather time regardless of whether we're in capture mode or
not.  It won't, of course, be valid when we don't force GL to
flush its pipeline, but in all modes we need to guarantee that the
method finishes in finite time.