Add FileBackedDataset class.
Derives from BrickedDataset, will fall in between BrickedDataset
and UVFDataset.

There are a few places in Tuvok where we need a filename: we try
to derive file names for default transfer function files, and then
there is the occasional debug message which wants to print the
name.  Our 'Dataset's, however, may not have an associated
filename.  Thus we attempt to upcast, and fallback to some reduced
behavior if the upcast fails.

This sucks for those attempting to implement support for formats
which are parallel to UVF, because they lose these benefits even
if they have an associated file.  FileBackedDataset will rectify
that; we can upcast to that, get our filename, and readers which
are parallel to UVF can derive from that class.