Fix the windows signedness warnings in the new minmax code.
At least on Windows, a unary negation operator cannot change the type of a
variable.  As such, our mechanism for computing a bogus initial value for
maximums spewed out a bunch of warnings.

This uses the type tag idiom (i.e. the idea used by iterators) to conditionalize
the initial max value on the signedness of the input type.