Add a 'Cleanup' method which essentially clones the destructor.
Since our MasterController is a singleton, it is initialized and
killed during static initialization/destruction.  One effect of
this is that tools like valgrind can get somewhat confused,
especially if you ask them to run earlier than process exit.

This method allows one to cleanup most of the memory that the
controller is holding at will.  Debug channels are held on to, in
case messages need to come out, but trying to do much else after
the controller has `cleaned up' will give undefined behavior.