Reimplement factory in terms of Datasets.
That is, don't require that we use FileBackedDatasets.  The
factory shouldn't have such knowledge.  I also added a `Reader'
method, which returns the instance which should be cloned.

We let that implementation creep in while searching for the
correct reader, because I wanted to leave `CanRead' in
FileBackedDataset (i.e. not move it up to Dataset).  Maybe we
could rename it `CanAccess'; right now the method is
file-specific, and the system works fine like that.

In the current system, if we create a new method of dataset access
which is not backed by a file, all we should need to do is modify
that one factory method, which seems like a good thing.

Also... "Reader" should really return a *set* of weak_ptr's --
multiple readers might be able to open the same file, or at least
think they can.  This is currently not a problem, but something to
keep in mind if a lot of users are keen on adding their own
formats.