mm: remove hack when searching for existing programs.
The interface was based on a single vertex and a single fragment
shader, and keyed into the filenames of those shaders to determine
if it could reuse the program.  This removes the lone shaders and
stores the full list of shaders used in a shader program, so that
we can correctly figure out if a program should be reused.