glsl: Remove shader program name strings.
We added these for better error messages.  Now, though, we can't
rely on being able to store the program filenames using just two
strings, and these strings ended up unset anyway.

We could store a vector of all the filenames, and use that for
error messages, but it makes formatting messages an ordeal and
isn't really that important anyway.  Maybe we'll go back to that
if we end up doing a lot of shader debugging, but that doesn't
seem as useful these days, and OpenGL debuggers work fairly well
anyway.