"Fix" loading Compose-Scanline-FS.glsl on Windows.
ifstream clearly tells us the file is 1838 bytes.  Explorer clearly thinks
the file is 1838 bytes.   Yet the read completes after 1700 and then some
bytes, and the EOF bit is set.  Further reads don't give any data.

Before complaining, we check for the EOF bit.  Further, we initialize the
text data to all 0's, instead of just the final character of the buffer,
to ensure that if we *do* see a short read, we'll get a NULL character and
terminate the string immediately afterwards.