Some fixes for reading text data.
When copying data, we would write it out before checking if the
read actually succeeded, in a loop.  We'd therefore always write
one extra element, because we didn't detect stream failure until
after doing the write.  The fix probably makes reading text data
slower... but reading text data is slow.  Don't use it.  We
probably never noticed this issue because we try to use the sizes
of data sets instead of the sizes of files.

32-bit g++-4.3.2 with optimizations would give a crazy warning
about potentially using an uninitialized value.  It seemed
impossible given the code, but by initializing to 0 we avoid the
issue entirely.

Finally, we checked for stream EOF in a loop.  However, if a
stream conversion failed, the failbit would get set but not the
eof bit, and no further reads would make progress.  This would
cause the loop to become infinite if the user put, say, a comment
character in the data.  I made the loop check for any failure
condition instead of just EOF.  If we fail to copy some data,
we'll notice later in the conversion process.