Add an assert to catch ridiculous brick sizes.
I hit this because I forgot a boolean argument to ConvertDataset.
Instead, my brick size argument was taken in boolean context, and
my brick overlap was taken as a max brick size.

Took me forever to track down that the reason I was generating 6
LoDs and 7000 bricks is because I specified a brick size of 4^3.
This assert should guard against future stupidities...