Guard against error states.
If we get into an error state, we can get INT_MAX for log_length.
This either comes about because a driver is braindead or log_length
happens to take that value because it was uninitialized; however it
comes about, a gigantic value causes the vector initialization to
take forever.