Don't always enable debug stl.
It will break at link time if libTuvok.a was not built with debug
stl as well.  By simply adding the flag during qmake-time, it
should cause /all/ projects to use the flag, and thus it ensures
we're consistent... at the risk of forgetting to enable it in some
cases.