Remove BIAS_SCALE from unlit path.
In most of our shaders we had a series of #ifdef's which handled
the case of unquantized data.  Such data must be biased and
scaled, not simply scaled, so that it falls within the range of
the transfer function.  Clients of the library in that fashion
were expected to open up these shaders and #define BIAS_SCALE all
over the place, which was a huge hassle and meant they had to
maintain a separate version of tuvok.

This commit removes BIAS_SCALE from the unlit path of tuvok.
Instead, GL-based renderers are expected to set a uniform,
'ScaleMethod', to indicate whether they want the traditional
scaling-only method (i.e. what IV3D wants) or the bias+scaling
method.

Doing this makes the interface to the volume rendering function
variable; there is an extra parameter required in the case of the
bias+scale path.  To combat this, there is a trampoline method
which has a uniform interface and simply reads GL uniforms plus
the 'ScaleMethod' to bounce into whichever scaling code is
appropriate.

A new initialization stage was added to the GL path, as well,
dubbed 'SetConstantShaderVars'.  This is for uniforms which
should stay the same over the life of a render window (or whatever
AbstrRenderer is tied to).  This fits our 'ScaleMethod' exactly;
the client code decides which method they want when a renderer is
created and it stays that way even if the underlying data changes.