Consistently use API for testing if we have color data.
Rename "RGBAData" to "ColorData", and use it throughout, instead
of manually checking component count. In the long term, we'll want
a more complicated check that queries the data set anyway. That is,
the data set might want to store 4-component data but say that it
isn't color data for whatever reason.