Rewrite a (hacky) special case for color data.
A rewrite/reorg a whiiiiile back now really broke multicomponent
data.  We were doing things like computing the max on a
multicomponent buffer as if it was a single component buffer.

We should rewrite that code to properly consider multicomponent
data... but we really should reorganize this code even more than
that.  There's a lot of cases and it's easy to mess things up,
as-is!

So, for now, just add a bit of a hack which separates out each
component, processes them appropriately, and then stitches
everything back together.  It's a bit slow, but it's only for
color data, and we'll get this fixed up soon enough.