dynbrick: Don't verify brick counts/LODs.
Rebricking stops when it becomes dumb to keep shrinking the
domain.  But our modern UVFs keep computing smaller LODs until
there is literally a 1x1x1 level of "detail".  So the number of
LODs and the number of bricks between a dataset and the equivalent
rebricked one won't be equivalent, since the original dataset will
have more LODs.

We could conceivably change this, but it's dumb and should work
without those LODs anyway, so instead let's not bother.