dynbrick: Need to translate the brick keys.
Lots of functions were just set up as forwarding functions, but
their input refers to the target dataset's bricks, and so if we
want to forward to the source dataset then we must first translate
the brick keys.

Abstract out the brick key computation so we can call it in one
line in a bunch of places.