lua: Provide a useful error message for unknown params.
If you try to bind a function with an argument type that we haven't
thought of yet, the compiler error message gives you all kind of
spew about ::Type being nonsense in the new type.

Took me forever to figure out why 'enum PerfCounter' couldn't be
used an argument: "of *course* it doesn't have a frickin' 'Type'
typedef, it's a basic type!!!"  It's not the basic type which
needs the typedef, however, but the LuaStrictStack template
specialization, which makes sure we only serialize/unserialize
things into lua explicitly.

So: now we actually check to make sure that overload actually
exists.  If no overload exists, the default templatized (invalid)
instantiation is chosen, which happens to give us 'void' for the
type.  Thus the 'Type' being void is equivalent to the StrictStack
specialization not being defined, and we now error out when that
happens.