Fix dynamic bricking cache for good.
Now we actually acquire exactly the amount of memory we think we do and
not more which prevents us from flooding our main memory.
Undo Tom's fix that was just a workaround hack based on the wrong
allocations.
The cache size uses 80% of the usable memory defined in the applications
settings.